+++ src/test/java/io/kiranhk/ebookstore/EbookStoreApplicationTests.java
@@ -1,10 +1,10 @@
-package io.kiranhk.library;
+package io.kiranhk.ebookstore;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class LibraryApplicationTests {
+class EbookStoreApplicationTests {
 
 	@Test
 	void contextLoads() {
